Various updates.

+++ Maestro.Common/Aircraft.cs
@@ -20,7 +20,7 @@ namespace Maestro.Common
         public string STAR { get; set; }
         public Position Position { get; set; }
         public int RFL { get; set; }
-        public int? Altitude { get; set; }
+        public double? Altitude { get; set; }
         public int? GroundSpeed { get; set; }
         public string Route { get; set; }
         public List<RoutePoint> RoutePoints { get; set; } = new List<RoutePoint>();
